Add comment-counter module and test
