+++ src/modules/commentsCounter.js
@@ -1,0 +1,9 @@
+const commentsCounter = (commentCount) => {
+  const commentCounter = document.querySelector('.comment-counter');
+
+  if (commentCounter) {
+    commentCounter.textContent = `Comments: ${commentCount}`;
+  }
+};
+
+export default commentsCounter;
